HOTFIX: Re-add cast to handle OS_SocketRecvFrom negative status

+++ fsw/src/ci_lab_app.c
@@ -365,7 +365,7 @@ void CI_LAB_ReadUpLink(void)
     {
         status = OS_SocketRecvFrom(CI_LAB_Global.SocketID, CI_LAB_Global.IngestBuffer.bytes,
                                    sizeof(CI_LAB_Global.IngestBuffer), &CI_LAB_Global.SocketAddress, OS_CHECK);
-        if (status >= sizeof(CFE_MSG_CommandHeader_t) && status <= ((int32)CI_LAB_MAX_INGEST))
+        if (status >= (int32)sizeof(CFE_MSG_CommandHeader_t) && status <= ((int32)CI_LAB_MAX_INGEST))
         {
             CFE_ES_PerfLogEntry(CI_LAB_SOCKET_RCV_PERF_ID);
             CI_LAB_Global.HkTlm.Payload.IngestPackets++;
